Fix loss of PU override properties
Memorize override PU props even if the required DataSource is not yet
available. Always merge the memorized props into the static PU props
before publishing an EntityManagerFactory.
We have hold on to the override props although we might not yet be able
to create an EntityManagerFactory, since we will not be notified about
the updated props again.

Resolves: ARIES-1992